Changed App working flow, it's now working as instructed:
The calculator should provide the following two modes of operation:

Menu: The application should display a menu in the console through which the user, via the keyboard, will provide all the information the program requires.

java Calculadora -m

Parameters: The user will indicate, through the command line, that they do not wish to use the calculator with a menu, specifying the operation they are requesting and the data for that operation.

java Calculadora -p -s 3 7

To correct:

	1.	In App.java, it should still allow direct command use by executing the command java -jar target/practica05-v0.0.1.jar -p -re 3.4, but if no arguments are provided, it should display a menu like:

Select mode of use:
-m -> menu
-p -> use of arguments

When using -m, it should display the menu as it has been doing from Menu.java. If -p is used, it should display -help and wait for the user’s input to pass the arguments to CommandControl.java.

	2.	In CommandControl.java, it should be modified so that it is no longer necessary to specify whether it is a scientific or basic operation. It should execute -p -re 2.56789 from App.java, passing the arguments to CommandControl.java, or execute java -jar target/practica05-v0.0.1.jar -p -re 3.4 without entering App.java.
	3.	In Menu.java, it should remain in a loop until the user selects “exit” to quit, and basic validations should be added so that it does not accept numbers or, if the option does not exist, shows an error and displays the options again.